Improved Sentis inference configuration
Summary:
Improved Sentis configuration to increase the accuracy:

* adjusted the IOU and SCORE values

* increased the camera image size to the maximum instead of 800x600

NOTE: the light in the room impacts directly into the model dection result. If is too dark the accuracy is lower. Also, this model is a tiny yolo 9 model, it means that the accuracy will be low.

NOTE 2: This is a pretrained model with just 80 classes. For objects that the model are not trained with, the result will be wrong (for example, Quest devices)

allow-large-files

Reviewed By: kyrylokuzyk-m

Differential Revision: D73423339

fbshipit-source-id: 695faaf94d41b0e793e332f416b6e2c8d183007a